backend/pyproject.toml
[tool.poetry]
name = "echoquest"
version = "0.1.0"
description = "Interview transcription and analysis platform"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.103.1"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
pydantic = {extras = ["email"], version = "^2.3.0"}
pydantic-settings = "^2.0.3"
sqlalchemy = "^2.0.20"
asyncpg = "^0.28.0"
alembic = "^1.12.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
httpx = "^0.24.1"
tenacity = "^8.2.3"
redis = "^5.0.0"
celery = "^5.3.4"
openai = "^0.28.1"
qdrant-client = "^1.6.0"
tiktoken = "^0.5.1"
bs4 = "^0.0.1"
pydub = "^0.25.1"
ffmpeg-python = "^0.2.0"
python-dotenv = "^1.0.0"
fastapi-pagination = "^0.12.10"
loguru = "^0.7.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
pytest-asyncio = "^0.21.1"
black = "^23.7.0"
isort = "^5.12.0"
mypy = "^1.5.1"
pylint = "^2.17.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"